Update gopsutil code
Latest gosutil includes two backward incompatible changes:

First, it removed unused Stolen field in
https://github.com/shirou/gopsutil/commit/cae8efcffa765c9534442328d385bdd883faf5df#diff-d9747e2da342bdb995f6389533ad1a3d
.

Second, it updated the Windows cpu stats calculation to be inline with
other platforms, where it returns absolate stats rather than
percentages.  See https://github.com/shirou/gopsutil/pull/611.
